requirements.txt
pandas==2.0.3
datasets==3.1.0
transformers==4.46.3
torch==2.4.1
accelerate==1.0.1
scikit-learn==1.3.2
google-cloud-storage==2.16.0